+++ docs/changelog.rst
@@ -7,6 +7,9 @@ v0.3 (not yet released)
 * Constructors: ``create_from_chain``
   `#28 <https://github.com/msoeken/kitty/pull/28>`_
 
+* Cube constructors: ``pos_cube``, ``neg_cube``
+  `#29 <https://github.com/msoeken/kitty/pull/29>`_
+
 v0.2 (December 21, 2017)
 ------------------------